+++ source/deprecated/omni.isaac.manipulators/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [3.0.7] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [3.0.6] - 2025-04-04
 ### Changed
 - Version bump to fix extension publishing issues