+++ source/extensions/isaacsim.core.deprecation_manager/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [0.2.7] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [0.2.6] - 2025-05-16
 ### Changed
 - Make extension target a specific kit version