+++ source/extensions/isaacsim.core.simulation_manager/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.0.1] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.0.0] - 2025-05-16
 ### Changed
 - Add time acquisition related features to this extension