+++ source/extensions/isaacsim.gui.property/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.3] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.1.2] - 2025-04-04
 ### Changed
 - Version bump to fix extension publishing issues