+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.6] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.1.5] - 2025-05-16
 ### Changed
 - Make extension target a specific kit version