+++ source/extensions/isaacsim.robot_setup.assembler/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [3.0.2] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [3.0.1] - 2025-05-16
 ### Changed
 - Make extension target a specific kit version