+++ source/extensions/isaacsim.ros2.urdf/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.14] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.1.13] - 2025-05-10
 ### Changed
 - Enable FSD in test settings