+++ source/extensions/isaacsim.sensors.camera/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.1] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.1.0] - 2025-05-13
 ### Added
 - attach_annotator and detach_annotator APIs to Camera