+++ source/extensions/isaacsim.simulation_app/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [2.8.1] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
 
 ## [2.8.0] - 2025-05-17
 ### Changed