+++ source/extensions/isaacsim.util.camera_inspector/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.0.10] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [1.0.9] - 2025-05-10
 ### Changed
 - Enable FSD in test settings