+++ source/extensions/isaacsim.util.merge_mesh/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.0.11] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [2.0.10] - 2025-05-16
 ### Changed
 - Make extension target a specific kit version