+++ source/extensions/omni.isaac.ml_archive/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [3.0.2] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
+
 ## [3.0.1] - 2025-04-10
 ### Changed
 - Update unit test