+++ source/extensions/omni.pip.compute/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.6.2] - 2025-05-19
+### Changed
+- Update copyright and license to apache v2.0
 
 ## [1.6.1] - 2025-05-12
 ### Changed